requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
attrs==22.1.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
decorator==5.1.1
distlib==0.3.4
Django==4.1.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
executing==0.8.3
filelock==3.7.1
fonttools==4.34.4
gunicorn==20.1.0
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
ipython==8.4.0
jedi==0.18.1
jsonschema==4.16.0
kiwisolver==1.4.3
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.23.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.4.2
stack-data==0.3.0
tinydb==4.7.0
toml==0.10.2
traitlets==5.3.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.15.1
wcwidth==0.2.5
zipp==3.8.1
